+++ resources/lang/en/admin/user.php
@@ -1,0 +1,20 @@
+<?php
+
+return [
+    'index' => [
+        'action' => 'Action',
+        'banned' => 'Bannned',
+        'count' => 'player(s)',
+        'details' => 'Details',
+        'email' => 'E-mail address',
+        'id' => 'ID',
+        'ip' => 'IP address',
+        'last_login_at' => 'Last login',
+        'no' => 'No',
+        'registered_at' => 'Registered at',
+        'table_title' => 'Registered players',
+        'title' => 'Players list',
+        'username' => 'Username',
+        'yes' => 'Yes',
+    ],
+];