+++ xnova/language/en/admin.php
@@ -9,21 +9,6 @@ $lang['Welcome_to_Fix_section'] = "Bienvenue dans la section d'update";
 $lang['There_is_not_need_fix']  = "La mise &agrave; jour a bien &eacute;t&eacute; effectu&eacute;!";
 $lang['Fix_welldone']           = "La mise &agrave; jour n'a pu &ecirc;tre effectu&eacute;.";
 
-$lang['adm_ul_title'] = "Players list";
-$lang['adm_ul_ttle2'] = "Registered players";
-$lang['adm_ul_id']    = "ID";
-$lang['adm_ul_name']  = "Username";
-$lang['adm_ul_mail']  = "E-mail address";
-$lang['adm_ul_adip']  = "IP address";
-$lang['adm_ul_regd']  = "Registered at";
-$lang['adm_ul_lconn'] = "Last login";
-$lang['adm_ul_bana']  = "Bannned";
-$lang['adm_ul_detai'] = "Details";
-$lang['adm_ul_actio'] = "Action";
-$lang['adm_ul_playe'] = " player(s)";
-$lang['adm_ul_yes']   = "Yes";
-$lang['adm_ul_no']    = "No";
-
 $lang['adm_pl_title'] = "Activit&eacute; des plan&egrave;tes";
 $lang['adm_pl_activ'] = "Plan&egrave;tes actives";
 $lang['adm_pl_name']  = "Noms des Plan&egrave;tes";